+++ c10/core/DispatchKey.cpp
@@ -1,4 +1,4 @@
-#include "c10/core/TensorTypeId.h"
+#include "c10/core/DispatchKey.h"
 
 namespace c10 {
